Authorization on controller (UserOwnPost & AdminOnly)

+++ app/Http/Controllers/UserPostController.php
@@ -55,8 +55,10 @@ class UserPostController extends Controller
         return redirect('/');
     }
 
-    public function edit(Post $post): View
+    public function edit(Post $post): View|RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post))
+            return redirect()->back();
         $categories = Category::all();
         return view(
             'user.posts.edit',
@@ -66,6 +68,8 @@ class UserPostController extends Controller
 
     public function update(Post $post): RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post))
+            return redirect()->back();
         $attributes = $this->validatePost($post);
 
         if ($attributes['thumbnail'] ?? false) {
@@ -79,6 +83,8 @@ class UserPostController extends Controller
 
     public function destroy(Post $post): RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post) && auth()->user()->cannot('admin'))
+            return redirect()->back();
         $post->delete();
         return redirect('/')->with('success', 'Successfully deleted post!');
     }
